fastmcp-agents-bridge/fastmcp_agents_bridge_pydantic_ai/pyproject.toml
[project]
name = "fastmcp-agents-bridge-pydantic-ai"
version = "0.1.4"
description = "Pydantic AI Agent Bridge"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp",
    "logfire>=3.25.0",
    "pydantic-ai>=0.4.4",
]

[dependency-groups]
dev = [
  "pytest",
  "pytest-mock",
  "pytest-asyncio",
  "ruff",
  "basedpyright>=1.30.1",
]

lint = [
  "ruff"
]

[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "fastmcp_agents.bridge.pydantic_ai"

[tool.uv.sources]
fastmcp = { git = "https://github.com/jlowin/fastmcp" }

[tool.ruff]
extend="../../pyproject.toml"

[tool.pyright]
extends = "../../pyproject.toml"